requirements.txt
Django==4.2.13
django_rest_framework==0.1.0
pandas==2.2.2